Enhances ASP.NET Core route processing
