+++ pkg/lang/csharp/plugin_expose.go
@@ -488,7 +488,7 @@ func (c controllerSpec) resolveRoutes() []gatewayRouteDefinition {
 	return routes
 }
 
-var optionalLastSegmentPattern = regexp.MustCompile(`(^|/)(\{[^?{}]+\?})(/$|$)`)
+var optionalLastSegmentPattern = regexp.MustCompile(`(?:^|/)(\{[^?{}]+\?})(/$|$)`)
 
 // stripOptionalLastSegment strips the last segment of a path if it contains an optional param
 func stripOptionalLastSegment(routeTemplate string) string {
@@ -657,128 +657,4 @@ func parseActionAttributes(method MethodDeclaration) []actionSpec {
 	}
 
 	return specs
-}
-
-var expressParamConversionPattern = regexp.MustCompile(`\{([^:}?]*):?[^}]*}`)
-
-// sanitizeConventionalPath converts ASP.NET Core conventional path parameters to Express syntax,
-// but does not perform validation to ensure that the supplied string is a valid ASP.net route.
-// As such, there's no expectation of correct output for invalid paths
-// Regexp constraints, complex segments and controller/action token replacement are not yet supported
-func sanitizeConventionalPath(path string) string {
-	path = sanitizeRegexConstraints(path)
-
-	// convert to longest possible proxy route when required
-	firstProxyParamIndex := findFirstProxyRouteIndicator(path)
-	if firstProxyParamIndex > -1 {
-		path = path[0:firstProxyParamIndex]
-		path = path[0:strings.LastIndex(path, "{")+1] + "rest*}"
-	}
-
-	path = mergeComplexSegments(path)
-
-	// convert path params to express syntax
-
-	path = expressParamConversionPattern.ReplaceAllString(path, ":$1")
-	return path
-}
-
-var areaTokenPattern = regexp.MustCompile(`(?i)\[area]`)
-var actionTokenPattern = regexp.MustCompile(`(?i)\[action]`)
-var controllerTokenPattern = regexp.MustCompile(`(?i)\[controller]`)
-
-// sanitizeAttributeBasedPath converts ASP.NET Core attribute-based route path parameters to Express syntax,
-// but does not perform validation to ensure that the supplied string is a valid ASP.NET Core route.
-// As such, there's no expectation of correct output for invalid paths
-// Regex constraints and complex segments are not yet supported
-func sanitizeAttributeBasedPath(path string, area string, controller string, action string) string {
-	path = sanitizeRegexConstraints(path)
-
-	// replace params such as {controller=Index}
-	specialParamFormat := `(?i)\{\s*%s\s*=\s*%s\s*}`
-	path = regexp.MustCompile(fmt.Sprintf(specialParamFormat, "area", area)).ReplaceAllString(path, area)
-	path = regexp.MustCompile(fmt.Sprintf(specialParamFormat, "controller", controller)).ReplaceAllString(path, controller)
-	path = regexp.MustCompile(fmt.Sprintf(specialParamFormat, "action", action)).ReplaceAllString(path, action)
-
-	// convert to longest possible proxy route when required
-	firstProxyParamIndex := findFirstProxyRouteIndicator(path)
-	if firstProxyParamIndex > -1 {
-		path = path[0:firstProxyParamIndex]
-		path = path[0:strings.LastIndex(path, "{")+1] + "rest*}"
-	}
-
-	path = mergeComplexSegments(path)
-
-	// convert path params to express syntax
-	path = expressParamConversionPattern.ReplaceAllString(path, ":$1")
-
-	// replace special tokens
-	path = areaTokenPattern.ReplaceAllString(path, area)
-	path = controllerTokenPattern.ReplaceAllString(path, controller)
-	path = actionTokenPattern.ReplaceAllString(path, action)
-	return path
-}
-
-// '*' is a literal in ASP.NET Core routes when not the first character in a route param: /path/* -> /path/*, /path/{*slug} -> /path/:rest*
-// Routes containing '*' literals are not supported by AWS API Gateway.
-var catchAllRegexp = regexp.MustCompile(`\{\*`)
-
-// TODO: rework handling of default params as they might not always indicate the need for a proxy route
-//
-//	see: https://learn.microsoft.com/en-us/aspnet/core/fundamentals/routing?view=aspnetcore-6.0#route-template-reference
-func findFirstProxyRouteIndicator(path string) int {
-	firstProxyParamIndex := -1
-	if firstCatchAll := catchAllRegexp.FindStringIndex(path); firstCatchAll != nil {
-		firstProxyParamIndex = firstCatchAll[0] + 1 // +1 to avoid stripping the opening "{"
-	}
-	firstDefaultParam := strings.Index(path, "=")
-	if firstDefaultParam > -1 || (firstProxyParamIndex == -1 && firstDefaultParam < firstProxyParamIndex) {
-		firstProxyParamIndex = firstDefaultParam
-	}
-	return firstProxyParamIndex
-}
-
-var regexParamStartPattern = regexp.MustCompile(`(?i)\{[^:}]+:regex\(`)
-var regexStartPattern = regexp.MustCompile(`(?i):regex\(`)
-var regexEndPattern = regexp.MustCompile(`[^}](})(?:[^}]|$)`)
-
-func sanitizeRegexConstraints(path string) string {
-	sanitized := path
-	pStart := regexParamStartPattern.FindStringIndex(sanitized)
-	for ; pStart != nil; pStart = regexParamStartPattern.FindStringIndex(sanitized) {
-		rStart := regexStartPattern.FindStringIndex(sanitized[pStart[0]:])
-		if rStart == nil {
-			continue
-		}
-		regexStart := pStart[0] + rStart[0]
-		regexParamEnd := -1
-		regexEndSubmatches := regexEndPattern.FindStringSubmatchIndex(sanitized[regexStart:]) // [0,1] = complete match, [2,3] = capture group 1
-		if regexEndSubmatches != nil {
-			if regexEndSubmatches[2] != -1 {
-				regexParamEnd = regexStart + regexEndSubmatches[2]
-			}
-		}
-		if regexParamEnd != -1 {
-			sanitized = sanitized[0:regexStart] + sanitized[regexParamEnd:]
-		} else {
-			return path // prevents infinite loop if the route has any incomplete/invalid regex constraints
-		}
-	}
-	return sanitized
-}
-
-var complexSegmentPattern = regexp.MustCompile(`(?P<complex>[^{}/]+(?:[^{}/]*\{[^{}/]+}[^{}/]*?)+?|{[^{}/]+}(?:[^{}/]+|\{[^{}/]+})+)(?:/|$|/$)`)
-
-// mergeComplexSegments replaces complex path segments with uniquely named parameterized segments.
-// e.g. /simple/{c}ompl{ex}/{com}{plex} -> /simple/{complex1}/{complex2}
-func mergeComplexSegments(path string) string {
-	i := 0
-	return complexSegmentPattern.ReplaceAllStringFunc(path, func(complex string) string {
-		i++
-		segment := fmt.Sprintf("{%s%d}", "complex", i)
-		if strings.HasSuffix(complex, "/") {
-			segment += "/"
-		}
-		return segment
-	})
 }
